refactor(runtime): give the archive types and tool name one owner each
Three leftovers from moving the capability, each a second source of truth:

- `ToolResultArchiveRecorder` stayed in the compaction contract, which now
  imports the capability back — a type cycle. It belongs with the capability:
  it is one of the three services, and its input spans both prune paths.
- `'ArchiveRead'` was spelled in both the tool and the constant guarding
  against duplicate bindings. If those drift the guard silently stops
  guarding, so the tool now owns its own name.
- `ToolResultArchiveReader` was still imported by the backend after the field
  that used it was removed.

`ARCHIVE_READ_TOOL_NAME` also stops being public — nothing outside the
runtime uses it.

Docblocks corrected where the prose outlived the migration it described: the
dedup guard is not a migration affordance any more, "no spelling" overclaimed
(a capability with failing readers is constructible; a forgotten tool is not),
and the session-scoping duty the ceiling carve-out leans on was unwritten.
