+++ packages/runtime/src/index.ts
@@ -412,8 +412,6 @@ export type {
   SynthesisCacheWriter,
   SynthesisCacheWriteInput,
   SynthesisCacheWriteResult,
-  ToolResultArchiveRecorder,
-  ToolResultArchiveRecorderInput,
   SemanticCompactBlockRecorder,
 } from './ai-sdk-compaction-contract.js';
 export { PiAgentBackend, normalizePiAgentFrame } from './pi-agent-backend.js';
@@ -425,12 +423,11 @@ export type {
 } from './pi-agent-backend.js';
 
 export { buildBuiltinTools } from './builtin-tools.js';
-export {
-  createToolResultArchiveCapability,
-  ARCHIVE_READ_TOOL_NAME,
-} from './tool-result-archive-capability.js';
+export { createToolResultArchiveCapability } from './tool-result-archive-capability.js';
 export type {
   ToolResultArchiveCapability,
+  ToolResultArchiveRecorder,
+  ToolResultArchiveRecorderInput,
   ToolResultArchiveServices,
 } from './tool-result-archive-capability.js';
 export { queryTavily } from './tavily-search.js';
