+++ packages/runtime/src/tool-result-archive-capability.ts
@@ -14,13 +14,30 @@
  * longer part of the question.
  */
 
-import { buildArchiveReadTool } from './archive-read-tool.js';
-import type { ToolResultArchiveRecorder } from './ai-sdk-compaction-contract.js';
+import { ARCHIVE_READ_TOOL_NAME, buildArchiveReadTool } from './archive-read-tool.js';
+import type { ActiveToolResultArchiveCandidate } from './active-tool-result-prune.js';
+import type { StaleToolResultArchiveCandidate } from './context-budget.js';
 import type { ToolResultArchiveReader } from './tool-result-archive.js';
 import type { ToolResultArchiveResourceReader } from './tool-result-archive-resource.js';
 import type { MakaTool } from './tool-runtime.js';
 
-export const ARCHIVE_READ_TOOL_NAME = 'ArchiveRead';
+export { ARCHIVE_READ_TOOL_NAME };
+
+/**
+ * What the writer is handed for one pruned body. The union spans both prune
+ * paths — a stale prior-turn result and an active current-turn one — because
+ * the archive is one authority over both.
+ */
+export type ToolResultArchiveRecorderInput = (
+  | StaleToolResultArchiveCandidate
+  | (ActiveToolResultArchiveCandidate & { runtimeEventId: string })
+) & {
+  sessionId: string;
+  bodySha256: string;
+};
+export type ToolResultArchiveRecorder = (
+  input: ToolResultArchiveRecorderInput,
+) => Promise<{ artifactId: string } | void> | { artifactId: string } | void;
 
 /** Host-owned storage for one archive authority. */
 export interface ToolResultArchiveServices {
@@ -28,14 +45,26 @@ export interface ToolResultArchiveServices {
   archiveToolResult: ToolResultArchiveRecorder;
   /** Replay hydration, addressed by the originating runtime event. */
   readToolResultArchive: ToolResultArchiveReader;
-  /** Ref-addressed read backing `ArchiveRead`; carries no runtime-event identity. */
+  /**
+   * Ref-addressed read backing `ArchiveRead`. A `maka://archive/...` ref carries
+   * only artifact id, hash and size — no runtime-event identity — so this read
+   * must never synthesize one to satisfy the stricter replay path above.
+   *
+   * It MUST reject a read whose `sessionId` does not own the artifact. The
+   * runtime passes down the invoking session's id but cannot verify ownership
+   * itself, so this is the only place the session boundary is enforced, and the
+   * argument that the decoder grants a child no reach it lacked rests on it.
+   */
   readArchivedToolResultResource: ToolResultArchiveResourceReader['readArchivedToolResultResource'];
 }
 
 /**
- * One indivisible archive authority. Constructible only through
- * {@link createToolResultArchiveCapability}, so a half-built capability has no
- * spelling.
+ * One indivisible archive authority. Build it through
+ * {@link createToolResultArchiveCapability}: the decoder is derived from the
+ * same services the writer uses, so "archives, but the placeholder names a tool
+ * nobody bound" has no spelling. A capability whose readers merely fail is of
+ * course still constructible — but that failure is legible to the model and the
+ * host, which is what the old missing-tool state never was.
  */
 export interface ToolResultArchiveCapability {
   readonly services: ToolResultArchiveServices;
@@ -55,8 +84,13 @@ export function createToolResultArchiveCapability(
 
 /**
  * Add the decoder to a session's tool set when — and only when — that session
- * archives. Idempotent, so a host that still binds `ArchiveRead` itself keeps
- * exactly one of it while the surfaces are being migrated.
+ * archives.
+ *
+ * No in-tree host binds `ArchiveRead` any more, so the dedup guard exists for
+ * tool sets this runtime does not own: an embedder's custom tool, or an agent
+ * definition that names it. Such a tool wins, which also means it — not this
+ * capability — is what the placeholder's ref will be handed to. Advertising two
+ * tools under one name would be worse.
  */
 export function bindToolResultArchiveDecoder(
   tools: readonly MakaTool[],
